sun4i: Tweak cpufreq limits
cpu-freq has 2 sets of limits, 1 called cpuinfo which indicates what the
minimum / maximum is the cpu can be configured to run at, and the 2nd set
indicated between which limits the governor should automatically scale.

This patch uses this by:
1) Setting the cpuinfo max to 1.5 GHz, allowing overclocking without needing
to recompile the kernel
2) Setting the governer max to 1GHz, so no overclock is done unless the
user changes the governer max through sysfs
3) Raising the governer min from 60 MHz, which is unworkably slow, to
400 MHz, which is the last speed at which the cpu runs at its lowest
voltage of 1.0 volt.

3) Not only makes the interactive feel of the device a lot better, it
likely also *saves* power since it will cause the cpu to reach an idle
state quicker, which will allow it to power down entire blocks, thus stopping
the constant (non speed dependent) leakage current of those blocks sooner,
while using aprox. the same amount of power to do the work (it uses more
power due to the higher clock, but does the work proportionally faster,
resulting in aprox. the same power use). This is the so called race to idle
trick.

While at it also fix the formatting of sun4i_cpufreq_init().

Signed-off-by: Hans de Goede <hdegoede@redhat.com>