sunxi-serial: Deal with sunxi 8250-s special busy interrupt
The sunxi uarts report interrupts with the iir register no interrupt bit
set, which causes a stuck interrupt. I can reproduce this 100% on my mini-x
with 1 GB of RAM.

When this happens this bit (bit 0) is combined with bit 1 and bit 2, this
is also done by Synopsys DesignWare 8250, and is called a busy interrupt
condition there. Likely the sunxi and the Synopsys DesignWare 8250-s have some
common ancestry.

This patch fixes the stuck interrupt condition by re-using the code for the
busy-interrupt from 8250_dw.c after this ttyS0 works reliable on my mini-x.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>