mach-sunxi: configure IRQ for ARM Cortex-A8 performance monitoring unit
This allows the use of performance monitoring unit for statistical
sampling based profiling (oprofile, perf).

The downside is that the performance unit in Cortex-A8 is affected
by erratum 628216, which causes it to miss interrupts in some cases
and disrupt profiling results. Some information in public access
about this erratum can be found here:
    http://www.freescale.com/files/32bit/doc/errata/IMX53CE.pdf

As a result, it may be preferable to keep using the standard
timer for general purpose profiling instead of the PMU cycle
counter. But some of the other performance events may be still
useful in practice (cache misses, ...).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>