sunxi-hdmi: Fix pixelclock conversion harder
The pixelclock in hz -> picoseconds -> hz conversions loosing precision
first fixed by "sunxi-hdmi: Fix some modes no longer working with X",
is still a problem with some modes :|

For example 130000000 becomes 130010000, and X does not work.

We can only do pixelclocks which are a divider of 1 - 15 of a multiple of
3 MHz. Combining this with EDID's granularity of 10 Khz, all clocks must have
been a multiple of 100 or 250 KHz before the conversion.

Therefor this patch changes the rounding in fb_videomode_pixclock_to_hdmi_pclk
from rounding to 10Khz to rounding to 50Khz, which should hopefullly be
enough to get rid of any inprecision introduced by the conversion process.

Reported-by: Daniel Veillard <veillard@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>