sun4i_wemac: Add support for the mii-phy being connected to PortD
This is possible with the A10s (and used on the wobo-i5), and indicated
by emac_used being 2 rather then 1.

While working on this I noticed that the port related defines in
sun4i_wemac.h are wrong. The port registers are 0x24 bytes apart, not 0x20,
correct the define for PD_CFG0, which this patch uses, and remove all the
others as they are unused anyways.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>