ASoC: rt5645: Add jack detection workaround for GPD Win
GPD Win requires jd_mode=3 for making the jack detection working.
But it's not enough: the device returns the jack state wrongly, in an
inverted way.

For correcting this behavior, a new platform data flag,
jd_invert_logic, is introduced.  When this flag is set, the jack plug
state is flipped, and it returns the plugged when the hardware says
its unplugged, and vice versa.

The quirk is set up based on DMI string match.

Signed-off-by: Takashi Iwai <tiwai@suse.de>